# packages/daemon/openapi.yaml
openapi: 3.0.3
info:
  title: orch8 API
  description: |
    REST + WebSocket API for the orch8 orchestration platform.

    ## Authentication

    Two authentication paths are supported on all endpoints (except `GET /api/health`):

    **Agent Authentication** — Provide both `X-Agent-Id` and `X-Project-Id` headers. The agent is validated against the database.

    **Admin Authentication** — Requests from localhost (`127.0.0.1`, `::1`) without agent headers are treated as admin. Admins bypass all permission checks.
  version: 0.0.0
  license:
    name: Private
servers:
  - url: http://localhost:3847
    description: Local daemon
...
components:
  securitySchemes:
    AgentAuth:
      type: apiKey
      in: header
      name: X-Agent-Id
      description: Agent ID header (must be paired with X-Project-Id)
    ProjectScope:
      type: apiKey
      in: header
      name: X-Project-Id
      description: Project ID scope header
    RunContext:
      type: apiKey
      in: header
      name: X-Run-Id
      description: Optional run ID for correlation
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          example: not_found
        message:
          type: string
          example: Task not found
    ValidationError:
      type: object
      required: [error, details]
      properties:
        error:
          type: string
          enum: [validation_error]
        details:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
              path:
                type: array
                items:
                  type: string
              message:
                type: string
...
    Agent:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        name:
          type: string
        role:
          $ref: "#/components/schemas/AgentRole"
        status:
          $ref: "#/components/schemas/AgentStatus"
        icon:
          type: string
        color:
          type: string
        model:
          type: string
        effort:
          type: string
          nullable: true
        maxTurns:
          type: integer
        allowedTools:
          type: array
          items:
            type: string
        heartbeatEnabled:
          type: boolean
        heartbeatIntervalSec:
          type: integer
        wakeOnAssignment:
          type: boolean
        wakeOnOnDemand:
          type: boolean
        wakeOnAutomation:
          type: boolean
        maxConcurrentRuns:
          type: integer
        canAssignTo:
          type: array
          items:
            type: string
        canCreateTasks:
          type: boolean
        canMoveTo:
          type: array
          items:
            $ref: "#/components/schemas/TaskColumn"
        systemPrompt:
          type: string
        promptTemplate:
          type: string
        bootstrapPromptTemplate:
          type: string
        instructionsFilePath:
          type: string
          nullable: true
        budgetLimitUsd:
          type: number
          nullable: true
        budgetSpentUsd:
          type: number
        adapterType:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateAgent:
      type: object
      required: [id, projectId, name, role]
      properties:
        id:
          type: string
          minLength: 1
          maxLength: 100
        projectId:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
          maxLength: 200
        role:
          $ref: "#/components/schemas/AgentRole"
        icon:
          type: string
        color:
          type: string
        model:
          type: string
        effort:
          type: string
        maxTurns:
          type: integer
          minimum: 1
        allowedTools:
          type: array
          items:
            type: string
        heartbeatEnabled:
          type: boolean
        heartbeatIntervalSec:
          type: integer
          minimum: 0
        wakeOnAssignment:
          type: boolean
        wakeOnOnDemand:
          type: boolean
        wakeOnAutomation:
          type: boolean
        maxConcurrentRuns:
          type: integer
          minimum: 1
        canAssignTo:
          type: array
          items:
            type: string
        canCreateTasks:
          type: boolean
        canMoveTo:
          type: array
          items:
            $ref: "#/components/schemas/TaskColumn"
        systemPrompt:
          type: string
        promptTemplate:
          type: string
        bootstrapPromptTemplate:
          type: string
        instructionsFilePath:
          type: string
        mcpTools:
          type: array
          items:
            type: string
        desiredSkills:
          type: array
          items:
            type: string
          description: Skill slugs this agent should use (resolved from projectSkills at runtime)
        adapterType:
          type: string
        adapterConfig:
          type: object
          additionalProperties: true
        envVars:
          type: object
          additionalProperties:
            type: string
        budgetLimitUsd:
          type: number
          minimum: 0
        autoPauseThreshold:
          type: integer
          minimum: 0
          maximum: 100
        maxConcurrentTasks:
          type: integer
          minimum: 1
        maxConcurrentSubagents:
          type: integer
          minimum: 0
        workingHours:
          type: string
    UpdateAgent:
      type: object
      description: Same fields as CreateAgent, all optional. id and projectId are not updatable. effort, instructionsFilePath, and budgetLimitUsd are nullable.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        role:
          $ref: "#/components/schemas/AgentRole"
        icon:
          type: string
        color:
          type: string
        model:
          type: string
        effort:
          type: string
          nullable: true
        maxTurns:
          type: integer
          minimum: 1
        allowedTools:
          type: array
          items:
            type: string
        heartbeatEnabled:
          type: boolean
        heartbeatIntervalSec:
          type: integer
          minimum: 0
        wakeOnAssignment:
          type: boolean
        wakeOnOnDemand:
          type: boolean
        wakeOnAutomation:
          type: boolean
        maxConcurrentRuns:
          type: integer
          minimum: 1
        canAssignTo:
          type: array
          items:
            type: string
        canCreateTasks:
          type: boolean
        canMoveTo:
          type: array
          items:
            $ref: "#/components/schemas/TaskColumn"
        systemPrompt:
          type: string
        promptTemplate:
          type: string
        bootstrapPromptTemplate:
          type: string
        instructionsFilePath:
          type: string
          nullable: true
        mcpTools:
          type: array
          items:
            type: string
        desiredSkills:
          type: array
          items:
            type: string
          description: Skill slugs this agent should use (resolved from projectSkills at runtime)
        adapterType:
          type: string
        adapterConfig:
          type: object
          additionalProperties: true
        envVars:
          type: object
          additionalProperties:
            type: string
        budgetLimitUsd:
          type: number
          minimum: 0
          nullable: true
        autoPauseThreshold:
          type: integer
          minimum: 0
          maximum: 100
          nullable: true
        maxConcurrentTasks:
          type: integer
          minimum: 1
        maxConcurrentSubagents:
          type: integer
          minimum: 0
        workingHours:
          type: string
          nullable: true
    CloneAgent:
      type: object
      required: [targetProjectId, newId]
      properties:
        targetProjectId:
          type: string
          minLength: 1
        newId:
          type: string
          minLength: 1
          maxLength: 100

    # --- Comment ---
...
    AddBundledAgents:
      type: object
      required: [projectId, agentIds]
      properties:
        projectId:
          type: string
          minLength: 1
        agentIds:
          type: array
          minItems: 1
          items:
            type: string
            minLength: 1

    # --- Run Events ---
...
    TaskColumn:
      type: string
      enum: [backlog, blocked, in_progress, review, verification, done]
...
    AgentRole:
      type: string
      enum: [cto, engineer, qa, researcher, planner, implementer, reviewer, verifier, referee, custom]
    AgentStatus:
      type: string
      enum: [active, paused, terminated]
...
paths:
...
  /api/agents:
    post:
      tags: [Agents]
      summary: Create agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgent"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "409":
          description: Agent ID already exists in project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      tags: [Agents]
      summary: List agents
      parameters:
        - name: projectId
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            $ref: "#/components/schemas/AgentRole"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/AgentStatus"
      responses:
        "200":
          description: Agent list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
  /api/agents/{id}:
    get:
      tags: [Agents]
      summary: Get agent
      description: Requires X-Project-Id header (agents have composite PK).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "400":
          description: Missing projectId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      tags: [Agents]
      summary: Update agent
      description: Requires X-Project-Id header.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgent"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/agents/{id}/pause:
    post:
      tags: [Agents]
      summary: Pause agent
      description: Requires X-Project-Id header.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: Paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "400":
          description: Missing projectId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Already paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/agents/{id}/resume:
    post:
      tags: [Agents]
      summary: Resume agent
      description: Requires X-Project-Id header.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "400":
          description: Missing projectId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Not paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/agents/{id}/wake:
    post:
      tags: [Agents]
      summary: Manual on-demand wakeup
      description: Enqueues an on_demand wakeup. Requires X-Project-Id header.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                taskId:
                  type: string
                reason:
                  type: string
      responses:
        "201":
          description: Wakeup enqueued
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  agentId:
                    type: string
                  projectId:
                    type: string
                  source:
                    type: string
                  status:
                    type: string
        "400":
          description: Missing projectId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Agent is paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/agents/{id}/clone:
    post:
      tags: [Agents]
      summary: Clone agent to another project
      description: Requires X-Project-Id header (source project).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CloneAgent"
      responses:
        "201":
          description: Cloned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "404":
          description: Source agent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: ID already exists in target project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ============================================================
  # Runs
  # ============================================================
...
  /api/bundled-agents/add:
    post:
      tags: [Bundled Agents]
      summary: Add bundled agents to a project
      description: |
        Copies bundled agent template directories into the project's
        `.orch8/`, syncs project skills, and creates agent rows. Agents that
        already exist in the project are skipped.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddBundledAgents"
      responses:
        "201":
          description: Agents created (may be empty if all were duplicates)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"